+++ src/main/java/com/example/quizApp/repo/teacher/TeacherRepo.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.example.quizApp.model.teacher.Teacher;
 
 @Repository
-public interface TeacherRepo extends JpaRepository<Teacher, Long> {
+public interface TeacherRepo extends JpaRepository<Teacher, Integer> {
 	void deleteByFirstname(String firstname);
 
 }
